+++ lib/index.ts
@@ -4,14 +4,15 @@
 
 import * as pkg from '../package.json';
 
-export {Value} from './src/value';
+export {ListSeparator, SassList} from './src/value/list';
 export {SassBoolean, sassFalse, sassTrue} from './src/value/boolean';
 export {SassColor} from './src/value/color';
-export {ListSeparator, SassList} from './src/value/list';
+export {SassFunction} from './src/value/function';
 export {SassMap} from './src/value/map';
-export {sassNull} from './src/value/null';
 export {SassNumber} from './src/value/number';
 export {SassString} from './src/value/string';
+export {Value} from './src/value';
+export {sassNull} from './src/value/null';
 
 export {Exception} from './src/exception';
 export {